Packages created for different activities and associated classes. Appropriate files moved and refactored. Imports optimized after refactoring.
