+++ app/src/main/java/com/healthyorg/android/healthyapp/MainActivity.kt
@@ -4,6 +4,12 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import com.healthyorg.android.healthyapp.MoodClasses.MoodActivity
+import com.healthyorg.android.healthyapp.foodactivityclasses.FoodActivity
+import com.healthyorg.android.healthyapp.goalClasses.GoalsActivity
+import com.healthyorg.android.healthyapp.sleepClasses.SleepActivity
+import com.healthyorg.android.healthyapp.weightClasses.WeightActivity
+import com.healthyorg.android.healthyapp.workoutClasses.WorkoutActivity
 
 class MainActivity : AppCompatActivity() {
