+++ app/src/main/java/com/healthyorg/android/healthyapp/database/GoalsDatabase.kt
@@ -2,10 +2,7 @@ package com.healthyorg.android.healthyapp.database
 
 import android.content.Context
 import androidx.room.*
-import com.healthyorg.android.healthyapp.Daily_Weight
-import com.healthyorg.android.healthyapp.Goal
-import com.healthyorg.android.healthyapp.database.WeightDao
-import com.healthyorg.android.healthyapp.database.WeightTypeConverters
+import com.healthyorg.android.healthyapp.goalClasses.Goal
 
 @Database(entities = [Goal::class], version = 2)
 abstract class GoalsDatabase : RoomDatabase() {
