+++ app/src/main/java/com/healthyorg/android/healthyapp/database/WeightDatabase.kt
@@ -2,9 +2,7 @@ package com.healthyorg.android.healthyapp.database
 
 import android.content.Context
 import androidx.room.*
-import com.healthyorg.android.healthyapp.Daily_Weight
-import com.healthyorg.android.healthyapp.database.WeightDao
-import com.healthyorg.android.healthyapp.database.WeightTypeConverters
+import com.healthyorg.android.healthyapp.weightClasses.Daily_Weight
 
 @Database(entities = [Daily_Weight::class], version = 1)
 @TypeConverters(WeightTypeConverters::class)
