+++ app/src/main/java/com/healthyorg/android/healthyapp/foodactivityclasses/FoodListFragment.kt
@@ -13,7 +13,6 @@ import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.healthyorg.android.healthyapp.R
-import com.healthyorg.android.healthyapp.Daily_Weight
 
 private const val TAG = "FoodListFragment"
