+++ app/src/main/java/com/healthyorg/android/healthyapp/goalClasses/Goal.kt
@@ -1,4 +1,4 @@
-package com.healthyorg.android.healthyapp
+package com.healthyorg.android.healthyapp.goalClasses
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
