+++ app/src/main/java/com/healthyorg/android/healthyapp/sleepClasses/DailySleepMood.kt
@@ -1,4 +1,4 @@
-package com.healthyorg.android.healthyapp
+package com.healthyorg.android.healthyapp.sleepClasses
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 import java.util.*
 
 @Entity
-data class Daily_Mood (
-    @ColumnInfo var feelings: String,
+data class DailySleepMood (
+    @ColumnInfo var hours: Double,
     @PrimaryKey var date: Date = Date())
