+++ app/src/main/java/com/healthyorg/android/healthyapp/weightClasses/Daily_Weight.kt
@@ -1,4 +1,4 @@
-package com.healthyorg.android.healthyapp
+package com.healthyorg.android.healthyapp.weightClasses
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
