+++ app/src/main/java/com/healthyorg/android/healthyapp/workoutClasses/Daily_Workout.kt
@@ -1,4 +1,4 @@
-package com.healthyorg.android.healthyapp
+package com.healthyorg.android.healthyapp.workoutClasses
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
